微调界面图标，以便容纳追踪前台窗口的功能
