+++ src/WindowDebugger/Localizations/en.toml
@@ -42,6 +42,9 @@ App.UI.SettingsButton = "Settings"
 App.UI.TopmostButton.Content = "Topmost"
 App.UI.TopmostButton.ToolTip.Description = "Toggle the topmost state of the window."
 App.UI.TopmostButton.ToolTip.Title = "Topmost"
+App.UI.TrackButton.Content = "Track"
+App.UI.TrackButton.ToolTip.Description = "After activating this button, whenever the foreground window changes, this list will automatically select the latest foreground window.\nRight-click this button to view the history of foreground windows during this activation period."
+App.UI.TrackButton.ToolTip.Title = "Track Foreground Window"
 App.UI.UacButton.Content = "UAC"
 App.UI.UacButton.ToolTip.Description = "Restart this program with administrator privileges so that you can view and modify information of high privilege windows."
 App.UI.UacButton.ToolTip.Title = "Restart as Administrator"
