+++ src/WindowDebugger/Localizations/zh-hans.toml
@@ -1,6 +1,6 @@
 App.UI.AboutButton = "关于"
 App.UI.CaptureButton.Content = "捕获"
-App.UI.CaptureButton.ToolTip.Description = "点击此按钮后，下一次鼠标点击的窗口将被捕获并在此列表中选中。"
+App.UI.CaptureButton.ToolTip.Description = "请拖动此按钮到目标窗口上方，目标窗口将被捕获并在此列表中选中。"
 App.UI.CaptureButton.ToolTip.Title = "捕获窗口"
 App.UI.Filter.FilterDescription = "默认在列表中显示顶层可见窗口，勾选以显示更多窗口。"
 App.UI.Filter.FilterTitle = "包含窗口"
@@ -42,6 +42,9 @@ App.UI.SettingsButton = "设置"
 App.UI.TopmostButton.Content = "置顶"
 App.UI.TopmostButton.ToolTip.Description = "将当前窗口置顶显示。"
 App.UI.TopmostButton.ToolTip.Title = "置顶"
+App.UI.TrackButton.Content = "追踪"
+App.UI.TrackButton.ToolTip.Description = "激活此按钮后，一旦前台窗口发生变化，此列表将自动选中最新的前台窗口。\n右键点击此按钮可查看本次按钮激活期间前台窗口的历史记录。"
+App.UI.TrackButton.ToolTip.Title = "追踪前台窗口"
 App.UI.UacButton.Content = "管理员"
 App.UI.UacButton.ToolTip.Description = "以管理员权限重新启动此程序，以便可以查看和修改高权限窗口的信息。"
 App.UI.UacButton.ToolTip.Title = "管理员权限"
