+++ src/WindowDebugger/Views/MainView.axaml
@@ -37,6 +37,21 @@
             <Setter Property="Height" Value="16" />
             <Setter Property="Stretch" Value="Uniform" />
         </Style>
+        <Style Selector="Svg.Reload">
+            <Setter Property="Path" Value="/Assets/UI/reload.svg" />
+        </Style>
+        <Style Selector="Svg.Capture">
+            <Setter Property="Path" Value="/Assets/UI/crosshair.svg" />
+        </Style>
+        <Style Selector="Button:pressed Svg.Capture">
+            <Setter Property="Path" Value="/Assets/UI/crosshair-fill.svg" />
+        </Style>
+        <Style Selector="Svg.Selection">
+            <Setter Property="Path" Value="/Assets/UI/selection.svg" />
+        </Style>
+        <Style Selector="ToggleButton:checked Svg.Selection">
+            <Setter Property="Path" Value="/Assets/UI/selection-fill.svg" />
+        </Style>
         <Style Selector="TextBlock.FlyoutTitle">
             <Setter Property="FontSize" Value="14" />
             <Setter Property="FontWeight" Value="Bold" />
@@ -112,7 +127,7 @@
         <!-- Left Side Bar -->
         <Grid Grid.Column="0" Margin="0 40 0 0" RowDefinitions="Auto,*,Auto">
             <StackPanel Grid.Row="0">
-                <Grid Margin="8 0" ColumnDefinitions="Auto,Auto,*,Auto,Auto">
+                <Grid Margin="8 0" ColumnDefinitions="Auto,*,Auto,Auto">
                     <Button Grid.Column="0" Classes="GlobalToolButton"
                             Padding="0 0 12 0" HorizontalContentAlignment="Stretch"
                             Background="#10FFFFFF"
@@ -125,13 +140,13 @@
                             </TextBlock>
                         </ToolTip.Tip>
                         <Grid ColumnDefinitions="32,Auto">
-                            <Svg Grid.Column="0" Classes="ButtonIcon" Path="/Assets/UI/reload-svgrepo-com.svg" />
+                            <Svg Grid.Column="0" Classes="ButtonIcon Reload" />
                             <TextBlock Grid.Column="1" VerticalAlignment="Center" Text="{Binding App.UI.ReloadButton.Content, Source={x:Static l:Lang.Current}}" />
                         </Grid>
                     </Button>
-                    <ToggleButton Grid.Column="1" Classes="GlobalToolButton"
+                    <Button Grid.Column="2" Classes="GlobalToolButton"
                                   Margin="4 0 0 0" HorizontalContentAlignment="Stretch"
-                                  IsCheckedChanged="CaptureButton_IsCheckedChanged">
+                                  Click="CaptureButton_Click">
                         <ToolTip.Tip>
                             <TextBlock LineHeight="24">
                                 <Run FontSize="14" FontWeight="Bold" Text="{Binding App.UI.CaptureButton.ToolTip.Title, Source={x:Static l:Lang.Current}}" />
@@ -139,30 +154,25 @@
                                 <Run Text="{Binding App.UI.CaptureButton.ToolTip.Description, Source={x:Static l:Lang.Current}}" />
                             </TextBlock>
                         </ToolTip.Tip>
-                        <Svg Classes="ButtonIcon" Path="/Assets/UI/visibility-view-svgrepo-com.svg" />
-                    </ToggleButton>
-                    <ToggleButton Grid.Column="3" Classes="GlobalToolButton"
-                                  IsCheckedChanged="TopMostToggleButton_IsCheckedChanged">
-                        <ToolTip.Tip>
-                            <TextBlock LineHeight="24">
-                                <Run FontSize="14" FontWeight="Bold" Text="{Binding App.UI.TopmostButton.ToolTip.Title, Source={x:Static l:Lang.Current}}" />
-                                <LineBreak />
-                                <Run Text="{Binding App.UI.TopmostButton.ToolTip.Description, Source={x:Static l:Lang.Current}}" />
-                            </TextBlock>
-                        </ToolTip.Tip>
-                        <Svg Classes="ButtonIcon" Path="/Assets/UI/layers-svgrepo-com.svg" />
-                    </ToggleButton>
-                    <Button Grid.Column="4" Classes="GlobalToolButton" Margin="4 0 0 0" Click="UacButton_Click"
-                            IsVisible="{OnPlatform False, Windows=True}">
-                        <ToolTip.Tip>
-                            <TextBlock LineHeight="24">
-                                <Run FontSize="14" FontWeight="Bold" Text="{Binding App.UI.UacButton.ToolTip.Title, Source={x:Static l:Lang.Current}}" />
-                                <LineBreak />
-                                <Run Text="{Binding App.UI.UacButton.ToolTip.Description, Source={x:Static l:Lang.Current}}" />
-                            </TextBlock>
-                        </ToolTip.Tip>
-                        <Svg Classes="ButtonIcon" Path="/Assets/UI/shield-antivirus-svgrepo-com.svg" />
+                        <Svg Classes="ButtonIcon Capture" />
                     </Button>
+                    <ToggleButton Grid.Column="3" Classes="GlobalToolButton"
+                            Margin="4 0 0 0" HorizontalContentAlignment="Stretch"
+                            IsCheckedChanged="TrackButton_CheckedChanged">
+                        <ToolTip.Tip>
+                            <TextBlock LineHeight="24">
+                                <Run FontSize="14" FontWeight="Bold" Text="{Binding App.UI.TrackButton.ToolTip.Title, Source={x:Static l:Lang.Current}}" />
+                                <LineBreak />
+                                <Run Text="{Binding App.UI.TrackButton.ToolTip.Description, Source={x:Static l:Lang.Current}}" />
+                            </TextBlock>
+                        </ToolTip.Tip>
+                        <Svg Classes="ButtonIcon Selection" />
+                        <ToggleButton.ContextFlyout>
+                            <MenuFlyout>
+                                <MenuItem Header="查看前台窗口历史记录" />
+                            </MenuFlyout>
+                        </ToggleButton.ContextFlyout>
+                    </ToggleButton>
                 </Grid>
                 <TextBox Margin="8"
                          Watermark="{Binding App.UI.SearchWatermark, Source={x:Static l:Lang.Current}}"
@@ -289,7 +299,7 @@
                         HorizontalContentAlignment="Left" VerticalContentAlignment="Center"
                         Click="SettingsButton_Click">
                     <StackPanel Orientation="Horizontal">
-                        <Svg Margin="4 0 8 0" Classes="ButtonIcon" Path="/Assets/UI/settings-gear-svgrepo-com.svg" />
+                        <Svg Margin="4 0 8 0" Classes="ButtonIcon" Path="/Assets/UI/settings.svg" />
                         <TextBlock VerticalAlignment="Center"
                                    Text="{Binding App.UI.SettingsButton, Source={x:Static l:Lang.Current}}" />
                     </StackPanel>
@@ -333,7 +343,7 @@
                         </TextBlock>
                     </ToolTip.Tip>
                     <Grid ColumnDefinitions="32,Auto">
-                        <Svg Grid.Column="0" Classes="ButtonIcon" Path="/Assets/UI/reload-svgrepo-com.svg" />
+                        <Svg Grid.Column="0" Classes="ButtonIcon" Path="/Assets/UI/reload.svg" />
                         <TextBlock Grid.Column="1" VerticalAlignment="Center" Text="{Binding App.UI.WindowDetail.ReloadButton.Content, Source={x:Static l:Lang.Current}}" />
                     </Grid>
                 </Button>
@@ -349,7 +359,7 @@
                         </TextBlock>
                     </ToolTip.Tip>
                     <Grid ColumnDefinitions="32,Auto">
-                        <Svg Grid.Column="0" Classes="ButtonIcon" Path="/Assets/UI/folder-svgrepo-com.svg" />
+                        <Svg Grid.Column="0" Classes="ButtonIcon" Path="/Assets/UI/folder.svg" />
                         <TextBlock Grid.Column="1" VerticalAlignment="Center" Text="{Binding App.UI.WindowDetail.OpenDirectory.Content, Source={x:Static l:Lang.Current}}" />
                     </Grid>
                 </Button>
